dfdf

+++ Go/src/ERP/Order.go
@@ -92,7 +92,7 @@ func NewOrder(session *JHttp.Session) {
 	}
 	st.OrderID = id
 	st.CreatTime = CurTime()
-	st.TotleMoney = st.OrderNum * st.Money
+	st.TotleMoney = Decimal(st.OrderNum * st.Money)
 	go setLastOrderDate(curMon)
 	////////////////添加状态///////////////////////////////
 	appendStatus(st, st.UserName, st.CreatTime, "创建订单", Status_New)
@@ -230,7 +230,7 @@ func ModOrder(session *JHttp.Session) {
 	data.CustomBatch = st.CustomBatch
 	data.CustomNote = st.CustomNote
 	data.OrderNum = st.OrderNum
-	data.TotleMoney = data.OrderNum * data.Money
+	data.TotleMoney = Decimal(data.OrderNum * data.Money)
 
 	////////////////添加状态///////////////////////////////
 	appendStatus(data, data.UserName, CurTime(), "修改订单", getStatus(data.OrderNum, data.ProduceNum, data.SuccessNum))
@@ -274,7 +274,7 @@ func ModOrderPrice(session *JHttp.Session) {
 	}
 
 	data.Money = st.Money
-	data.TotleMoney = data.OrderNum * st.Money
+	data.TotleMoney = Decimal(data.OrderNum * st.Money)
 	////////////////添加状态///////////////////////////////
 	appendStatus(data, data.UserName, CurTime(), "修改订单价格", getStatus(data.OrderNum, data.ProduceNum, data.SuccessNum))
